==> code/00_pipeline/01a_fastq_concatenation.ipynb <==
import os, subprocess

from tqdm import tqdm

#### Create folder outputs in original folder

september_batch = 'HLTCTAFX3'
september_folder = f'/media/UTELifeNAS/homes/Savvy/piQTL/DATA/bioinfo.iric.ca/seq/2a6890f284e7a257013e61cd6f6fba46/{september_batch}/fastq/'

final_folder = f'/media/UTELifeNAS/homes/Savvy/piQTL/DATA/original'
folder_names = [ folder for folder in os.listdir(september_folder) ]
create_folders = [ os.mkdir(os.path.join(final_folder, folder)) for folder in folder_names ]

#### Copy fastq in original folder from september and october batch folder
- Will generate the input file for parallel input file

september_batch = 'HLTCTAFX3'
october_batch = 'HWF2LAFX3'

september_folder = f'/media/UTELifeNAS/homes/Savvy/piQTL/DATA/bioinfo.iric.ca/seq/2a6890f284e7a257013e61cd6f6fba46/{september_batch}/fastq/'
october_folder = f'/media/UTELifeNAS/homes/Savvy/piQTL/DATA/bioinfo.iric.ca/seq/2a6890f284e7a257013e61cd6f6fba46/{october_batch}/fastq/'

f = open('./parallel_lines/group_fastq_parallel.txt', 'w')
for folder in tqdm(os.listdir(september_folder)):
    source1_folder = os.path.join(september_folder, folder)
    source2_folder = os.path.join(october_folder, folder)
    target_folder = os.path.join(final_folder, folder)
    #### Copy the fastq files
    for source_folder in [source1_folder, source2_folder]:
        f.write(f'cp {source_folder}/*.gz {target_folder}\n')
f.close()

#### To run in the terminal in your current working directory
! time parallel -j 4 < ./parallel_lines/group_fastq_parallel.txt

#### Merge the fastq within each sample folder

f = open('./parallel_lines/merge_fastq_parallel.txt', 'w')
for folder in tqdm(os.listdir(final_folder)):
    for R in ['R1', 'R2']:
        input_file = os.path.join(final_folder, folder, f'*_{R}_001.fastq.gz')
        output_file = os.path.join(final_folder, folder, f'{folder.replace("Sample_","")}_{R}.fastq.gz ')
        merge_cmd = f'cat {input_file} > {output_file}\n'
        f.write(merge_cmd)
f.close()

#### To run in the terminal in your current working directory
! time parallel -j 4 < ./parallel_lines/merge_fastq_parallel.txt
