==> code/00_pipeline/04a_inputs_for_QTL_plots.ipynb <==
import cv2, os 
import numpy as np 
import pandas as pd
import seaborn as sns
import matplotlib as mpl
import scipy.stats as stats
import matplotlib.pyplot as plt

from assocplots.manhattan import *
from assocplots.qqplot import *
from tqdm import tqdm

#### SNPs annotations
snps_loc = pd.read_csv('../../data/genotype_information/snps_annotations_genome-version-3-64-1.txt').rename(columns={'snp_id':'SNP'})
snps_loc['SNP'] = snps_loc['SNP'].astype(int)
snps_loc['SNP_ID'] = [ f'{snps_loc["chrom"][idx].replace("_","")}_{snps_loc["SNP"][idx]}' for idx in snps_loc.index]
snps_loc['position'] = snps_loc['position'].astype(int)

#### Gene names 
gene_loc = pd.read_csv('../../data/genotype_information/yeast_ORFs_dec2022.txt')

#### Genotype matrix
genotype = pd.read_csv('../../data/genotype_information/piQTL_genotype_matrix_dec2022.txt')
nb_strains = len(genotype.columns) - 1

genotype

gene_loc

def get_chr(x, gene_loc):
    try : 
        return gene_loc[gene_loc['name'] == x]['chrom'].values[0].replace('CHR_', '')
    except : 
        return np.nan

PPI_list = pd.read_csv('../../data/pipeline/PPI_reference_barcodes.csv')
PPI_list['PPI_bait'] =  [ PPI_list['PPI'][idx].split('_')[0] for idx in PPI_list.index ]
PPI_list['PPI_prey'] =  [ PPI_list['PPI'][idx].split('_')[1] for idx in PPI_list.index ]

PPI_list['CHR_bait'] = PPI_list['PPI_bait'].apply(get_chr, args=(gene_loc,))
PPI_list['CHR_prey'] = PPI_list['PPI_prey'].apply(get_chr, args=(gene_loc,))

PPI_list

snps_loc

input_path = f'../../results/04a_rMVP/rMVP_outputs/piQTL_dec_2022/'
output_path = f'../../results/04a_rMVP/rMVP_outputs/piQTL_dec_2022_annotated/'
for table in tqdm(os.listdir(os.path.join(input_path))) :
    qtl_res = pd.read_csv(os.path.join(input_path, table))
    qtl_res = qtl_res.merge(snps_loc[['SNP', 'locus_id', 'name','sgd_id', 'snps_class_up', 'snps_class_down', 'description',  'genome_annotations']], on='SNP')
    qtl_res.to_csv(os.path.join(output_path, table), index=False)

! cp ../../results/04a_rMVP/rMVP_outputs/piQTL_dec_2022_annotated/*.csv /home/savvy/PROJECTS/PHD/piQTL/data/QTL

! cp ../../results/04a_rMVP/rMVP_outputs/piQTL_dec_2022_annotated/*.csv /home/savvy/PROJECTS/PHD/piQTL/data/QTL_v2

folder_path = '/home/savvy/PROJECTS/PHD/piQTL/data/QTL'
for PPI in tqdm(PPI_list['PPI']): 
    for DRUG in ['noDrug', '5.FC', 'Fluconazole', 'Metformin', 'Trifluoperazine'] : 
        # print(PPI)
        CHR_bait =  PPI_list[PPI_list['PPI'] == PPI]['CHR_bait'].values[0] 
        CHR_prey = PPI_list[PPI_list['PPI'] == PPI]['CHR_prey'].values[0] 
        # print(CHR_prey, CHR_bait)

        MTX = pd.read_csv(os.path.join(folder_path,f'{PPI}_MTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv')).replace('MT', 17)
        MTX = noMTX.rename(columns={'Chr':'CHR', 'Pos':'BP', f'{PPI}_MTX_{DRUG}_avg_logratio_Fitness_minus_ref.GLM':'P', 'Effect':'EFFECTSIZE', 'name':'GENE'})
        # MTX.to_csv(os.path.join(folder_path,f'{PPI}_MTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv'), index=False)
        
        noMTX = pd.read_csv(os.path.join(folder_path,f'{PPI}_noMTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv')).replace('MT', 17)
        noMTX = noMTX.rename(columns={'Chr':'CHR', 'Pos':'BP', f'{PPI}_noMTX_{DRUG}_avg_logratio_Fitness_minus_ref.GLM':'P', 'Effect':'EFFECTSIZE', 'name':'GENE'})
        # noMTX.to_csv(os.path.join(folder_path,f'{PPI}_noMTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv'), index=False)

        

eQTL_res = pd.read_csv('../../results/04b_eQTL_matrix/piQTL_dec2022_pval_results.txt', sep='\t')

folder_path = '/home/savvy/PROJECTS/PHD/piQTL/data/QTL_v2'
for PPI in tqdm(PPI_list['PPI']): 
    for DRUG in ['noDrug', '5.FC', 'Fluconazole', 'Metformin', 'Trifluoperazine'] : 
        # print(PPI)
        CHR_bait =  PPI_list[PPI_list['PPI'] == PPI]['CHR_bait'].values[0] 
        CHR_prey = PPI_list[PPI_list['PPI'] == PPI]['CHR_prey'].values[0] 
        # print(CHR_prey, CHR_bait)

        MTX = pd.read_csv(os.path.join(folder_path,f'{PPI}_MTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv')).replace('MT', 17)
        eQTL_MTX_RES = eQTL_res[(eQTL_res['gene'].str.contains(PPI)) & (eQTL_res['gene'].str.contains(DRUG)) & (eQTL_res['gene'].str.contains('_MTX'))]
        MTX = eQTL_MTX_RES.merge(MTX, on='SNP')
        MTX = MTX.rename(columns={'Chr':'CHR', 'Pos':'BP', 'p-value':'P', 'beta':'EFFECTSIZE', 'name':'GENE'})
        MTX.to_csv(os.path.join(folder_path,f'{PPI}_MTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv'), index=False)
        
        noMTX = pd.read_csv(os.path.join(folder_path,f'{PPI}_noMTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv')).replace('MT', 17)
        eQTL_noMTX_RES = eQTL_res[(eQTL_res['gene'].str.contains(PPI)) & (eQTL_res['gene'].str.contains(DRUG)) & (eQTL_res['gene'].str.contains('_noMTX'))]
        noMTX = eQTL_noMTX_RES.merge(noMTX, on='SNP')
        noMTX = MTX.rename(columns={'Chr':'CHR', 'Pos':'BP', 'p-value':'P', 'beta':'EFFECTSIZE', 'name':'GENE'})  
        noMTX.to_csv(os.path.join(folder_path,f'{PPI}_noMTX_{DRUG}_avg_logratio_Fitness_minus_ref.csv'), index=False)
